+++ modules/apps/headless/headless-foundation/headless-foundation-impl/src/main/java/com/liferay/headless/foundation/internal/resource/BaseWebUrlResourceImpl.java
@@ -31,7 +31,7 @@ import javax.annotation.Generated;
 public abstract class BaseWebUrlResourceImpl implements WebUrlResource {
 
 	@Override
-	public WebUrl getWebUrl(Integer id) throws Exception {
+	public WebUrl getWebUrl(Long id) throws Exception {
 		return new WebUrl();
 	}
