+++ modules/apps/headless/headless-web-experience/headless-web-experience-impl/src/main/java/com/liferay/headless/web/experience/internal/resource/BaseAggregateRatingResourceImpl.java
@@ -28,7 +28,7 @@ public abstract class BaseAggregateRatingResourceImpl
 	implements AggregateRatingResource {
 
 	@Override
-	public AggregateRating getAggregateRating(Integer id) throws Exception {
+	public AggregateRating getAggregateRating(Long id) throws Exception {
 		return new AggregateRating();
 	}
