+++ modules/apps/headless/headless-web-experience/headless-web-experience-impl/src/main/java/com/liferay/headless/web/experience/internal/resource/BaseContentDocumentResourceImpl.java
@@ -28,7 +28,7 @@ public abstract class BaseContentDocumentResourceImpl
 	implements ContentDocumentResource {
 
 	@Override
-	public ContentDocument getContentDocument(Integer id) throws Exception {
+	public ContentDocument getContentDocument(Long id) throws Exception {
 		return new ContentDocument();
 	}
